// packages/sdk-platform/src/generated-admin/endpoints/InvoiceAdmin$.ts
/* 
 * Copyright (c) 2022-2024 AccelByte Inc. All Rights Reserved
 * This is licensed software from AccelByte Inc, for limitations
 * and restrictions contact your company contract manager.
 */
/**
 * AUTO GENERATED
 */
import { Response, Validate } from '@accelbyte/sdk'
import { AxiosInstance, AxiosRequestConfig } from 'axios'
import { z } from 'zod'
import { InvoiceSummary } from '../../generated-definitions/InvoiceSummary.js'

export class InvoiceAdmin$ {
  // @ts-ignore
  // prettier-ignore
  constructor(private axiosInstance: AxiosInstance, private namespace: string, private useSchemaValidation = true) {}
  
  /**
   * Generate invoice summary.&lt;br&gt;Other detail info: &lt;ul&gt;&lt;li&gt;&lt;i&gt;Returns&lt;/i&gt;: query orders&lt;/li&gt;&lt;/ul&gt; 
   */
  getInvoiceSummary( queryParams: {endTime: string | null, feature: string | null, itemId: string | null, itemType: 'APP' | 'BUNDLE' | 'CODE' | 'COINS' | 'EXTENSION' | 'INGAMEITEM' | 'LOOTBOX' | 'MEDIA' | 'OPTIONBOX' | 'SEASON' | 'SUBSCRIPTION', startTime: string | null}): Promise<Response<InvoiceSummary>> {
    const params = { ...queryParams} as AxiosRequestConfig
    const url = '/platform/admin/namespaces/{namespace}/invoice/summary'.replace('{namespace}', this.namespace)     
    const resultPromise = this.axiosInstance.get(url, {params})

    return Validate.validateOrReturnResponse(this.useSchemaValidation, () => resultPromise, InvoiceSummary, 'InvoiceSummary')
  }
  
  /**
   * Download invoice details as a csv file.&lt;br&gt;Other detail info: &lt;ul&gt;&lt;li&gt;&lt;i&gt;Returns&lt;/i&gt;: invoice details csv file&lt;/li&gt;&lt;/ul&gt; 
   */
  getInvoiceDetailsCsv( queryParams: {endTime: string | null, feature: string | null, itemId: string | null, itemType: 'APP' | 'BUNDLE' | 'CODE' | 'COINS' | 'EXTENSION' | 'INGAMEITEM' | 'LOOTBOX' | 'MEDIA' | 'OPTIONBOX' | 'SEASON' | 'SUBSCRIPTION', startTime: string | null}): Promise<Response<unknown>> {
    const params = { ...queryParams} as AxiosRequestConfig
    const url = '/platform/admin/namespaces/{namespace}/invoice/details.csv'.replace('{namespace}', this.namespace)     
    const resultPromise = this.axiosInstance.get(url, {params})

    return Validate.validateOrReturnResponse(this.useSchemaValidation, () => resultPromise, z.unknown(), 'z.unknown()')
  }
  
}
  